Restructure cli
