+++ cmd/helm-schema-gen/cli/generate_test.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"path/filepath"
