+++ cmd/helm-schema-gen/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/cpanato/helm-schema-gen/cmd/helm-schema-gen/cli"
+
+func main() {
+	cli.Execute()
+}
