--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/cpanato/helm-schema-gen/cmd"
-
-func main() {
-	cmd.Execute()
-}
